IDNA Fixes; #627

+++ webpp/unicode/details/ccc_tables.hpp
@@ -6,7 +6,7 @@
  *
  *   Auto generated from:                generate_ccc_tables.mjs
  *   Unicode UCD Database Creation Date: 2024-08-25
- *   This file's generation date:        Tue, 22 Apr 2025 20:36:20 GMT
+ *   This file's generation date:        Wed, 23 Apr 2025 20:08:30 GMT
  *   Unicode Version:                    16.0.0
  *   Total Table sizes in this file:
  *       - in bits:       82824
@@ -134,7 +134,7 @@ namespace webpp::unicode::details {
           pos{static_cast<std::uint16_t>(value)} {}
 
         [[nodiscard]] constexpr std::uint16_t value() const noexcept {
-            return static_cast<std::uint16_t>(pos);
+            return static_cast<std::uint16_t>(static_cast<std::uint16_t>(pos));
         }
 
         static constexpr std::uint16_t chunk_mask  = 0x7FU;
@@ -143,7 +143,7 @@ namespace webpp::unicode::details {
 
         /**
          * Get the final position of the second table.
-         * This does not apply the shift or get the value of the second table for you; this only applies tha
+         * This does not apply the shift or get the value of the second table for you; this only applies that
          * mask.
          */
         [[nodiscard]] constexpr std::uint16_t get_position(auto const request_position) const noexcept {
